+++ src/Twilio/Rest/Video/V1/RoomResource.cs
@@ -191,6 +191,7 @@ namespace Twilio.Rest.Video.V1
         /// <param name="emptyRoomTimeout"> Configures the time a room will remain active after last participant leaves.
         ///                        </param>
         /// <param name="unusedRoomTimeout"> Configures the time a room will remain active when no one joins. </param>
+        /// <param name="largeRoom"> Indicates whether this is a large room. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> A single instance of Room </returns>
         public static RoomResource Create(bool? enableTurn = null,
@@ -207,9 +208,10 @@ namespace Twilio.Rest.Video.V1
                                           int? maxParticipantDuration = null,
                                           int? emptyRoomTimeout = null,
                                           int? unusedRoomTimeout = null,
+                                          bool? largeRoom = null,
                                           ITwilioRestClient client = null)
         {
-            var options = new CreateRoomOptions(){EnableTurn = enableTurn, Type = type, UniqueName = uniqueName, StatusCallback = statusCallback, StatusCallbackMethod = statusCallbackMethod, MaxParticipants = maxParticipants, RecordParticipantsOnConnect = recordParticipantsOnConnect, VideoCodecs = videoCodecs, MediaRegion = mediaRegion, RecordingRules = recordingRules, AudioOnly = audioOnly, MaxParticipantDuration = maxParticipantDuration, EmptyRoomTimeout = emptyRoomTimeout, UnusedRoomTimeout = unusedRoomTimeout};
+            var options = new CreateRoomOptions(){EnableTurn = enableTurn, Type = type, UniqueName = uniqueName, StatusCallback = statusCallback, StatusCallbackMethod = statusCallbackMethod, MaxParticipants = maxParticipants, RecordParticipantsOnConnect = recordParticipantsOnConnect, VideoCodecs = videoCodecs, MediaRegion = mediaRegion, RecordingRules = recordingRules, AudioOnly = audioOnly, MaxParticipantDuration = maxParticipantDuration, EmptyRoomTimeout = emptyRoomTimeout, UnusedRoomTimeout = unusedRoomTimeout, LargeRoom = largeRoom};
             return Create(options, client);
         }
 
@@ -235,6 +237,7 @@ namespace Twilio.Rest.Video.V1
         /// <param name="emptyRoomTimeout"> Configures the time a room will remain active after last participant leaves.
         ///                        </param>
         /// <param name="unusedRoomTimeout"> Configures the time a room will remain active when no one joins. </param>
+        /// <param name="largeRoom"> Indicates whether this is a large room. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> Task that resolves to A single instance of Room </returns>
         public static async System.Threading.Tasks.Task<RoomResource> CreateAsync(bool? enableTurn = null,
@@ -251,9 +254,10 @@ namespace Twilio.Rest.Video.V1
                                                                                   int? maxParticipantDuration = null,
                                                                                   int? emptyRoomTimeout = null,
                                                                                   int? unusedRoomTimeout = null,
+                                                                                  bool? largeRoom = null,
                                                                                   ITwilioRestClient client = null)
         {
-            var options = new CreateRoomOptions(){EnableTurn = enableTurn, Type = type, UniqueName = uniqueName, StatusCallback = statusCallback, StatusCallbackMethod = statusCallbackMethod, MaxParticipants = maxParticipants, RecordParticipantsOnConnect = recordParticipantsOnConnect, VideoCodecs = videoCodecs, MediaRegion = mediaRegion, RecordingRules = recordingRules, AudioOnly = audioOnly, MaxParticipantDuration = maxParticipantDuration, EmptyRoomTimeout = emptyRoomTimeout, UnusedRoomTimeout = unusedRoomTimeout};
+            var options = new CreateRoomOptions(){EnableTurn = enableTurn, Type = type, UniqueName = uniqueName, StatusCallback = statusCallback, StatusCallbackMethod = statusCallbackMethod, MaxParticipants = maxParticipants, RecordParticipantsOnConnect = recordParticipantsOnConnect, VideoCodecs = videoCodecs, MediaRegion = mediaRegion, RecordingRules = recordingRules, AudioOnly = audioOnly, MaxParticipantDuration = maxParticipantDuration, EmptyRoomTimeout = emptyRoomTimeout, UnusedRoomTimeout = unusedRoomTimeout, LargeRoom = largeRoom};
             return await CreateAsync(options, client);
         }
         #endif
@@ -602,6 +606,11 @@ namespace Twilio.Rest.Video.V1
         /// </summary>
         [JsonProperty("unused_room_timeout")]
         public int? UnusedRoomTimeout { get; private set; }
+        /// <summary>
+        /// Indicates if this is a large room.
+        /// </summary>
+        [JsonProperty("large_room")]
+        public bool? LargeRoom { get; private set; }
         /// <summary>
         /// The absolute URL of the resource
         /// </summary>